calcolaPerimetroArea cambio var vers1.2

+++ es_06/calcolaPerimetroArea.java
@@ -1,7 +1,7 @@
 
 public class calcolaPerimetroArea { //la classe deve corrispondere al nome del file
     public static void main(String[] args) { 
-        //le istruzioni vanno scritte all'interno delle graffe del metodo main
-        int lato; // questo sarà il parametro che utilizzerò per i calcoli
+
+        double lato; // questo sarà il parametro che utilizzerò per i calcoli
     }
 }
